# xswagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LAARTESA API",
    "version": "1.0.0",
    "description": "Documentación de la API para WEB APP LA ARTESA"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    },
    {
      "url": "https://f686-160-173-184-21.ngrok-free.app",
      "description": "Servidor Ngrok"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingrese el token JWT en el formato: Bearer <token>"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/password/request-reset": {
      "post": {
        "tags": ["Auth"],
        "summary": "Solicitar recuperación de contraseña",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mail"],
                "properties": {
                  "mail": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud procesada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Si el correo existe, recibirás instrucciones para restablecer tu contraseña"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/password/reset": {
      "post": {
        "tags": ["Auth"],
        "summary": "Restablecer contraseña usando token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "newPassword"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Token recibido por correo electrónico"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "minLength": 6,
                    "description": "Nueva contraseña del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contraseña actualizada exitosamente"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token inválido o expirado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  }
}